requirements.txt
pymupdf>=1.27.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pytest>=8.0.0
